FOGL-2544: splitted files update
FOGL-2544: splitted files update

MAX_RETRIES warnings has to be fixed
